Add: Custom hoook - useSkills
